refactor(Editor): move all codemirror plugins in plugins/codemirror
