+++ src/components/inputs/Codemirror.vue
@@ -14,57 +14,20 @@ import { basicSetup } from 'codemirror'
 import { EditorView, keymap } from '@codemirror/view'
 import { EditorState, Prec } from '@codemirror/state'
 import { vscodeDark, vscodeLight } from '@uiw/codemirror-theme-vscode'
-import { HighlightStyle, indentUnit, syntaxHighlighting } from '@codemirror/language'
-import { klipperConfig, gcode, klipperConfigLanguage } from '@/plugins/lezer'
-import { KlipperDocsTooltip } from '@/plugins/KlipperDocsTooltip'
+import { indentUnit, syntaxHighlighting } from '@codemirror/language'
+import { klipperConfig, gcode } from '@/plugins/codemirror/lezer'
+import { KlipperDocsTooltip } from '@/plugins/codemirror/KlipperDocsTooltip'
 import { indentLess } from '@codemirror/commands'
 import { json } from '@codemirror/lang-json'
 import { css } from '@codemirror/lang-css'
-import { yaml, yamlLanguage } from '@codemirror/lang-yaml'
-import { tags } from '@lezer/highlight'
+import { yaml } from '@codemirror/lang-yaml'
 import { insertSmartTab } from '@/plugins/codemirror/insertSmartTab'
-
-const yamlDarkHighlightStyle = HighlightStyle.define(
-    [
-        {
-            tag: tags.definition(tags.propertyName),
-            color: '#dcdcaa',
-        },
-    ],
-    { scope: yamlLanguage, themeType: 'dark' }
-)
-
-const yamlLightHighlightStyle = HighlightStyle.define(
-    [
-        {
-            tag: tags.definition(tags.propertyName),
-            color: '#795e26',
-        },
-    ],
-    { scope: yamlLanguage, themeType: 'light' }
-)
-
-// config keys (description:, gcode:, ...) in the darker vscode keyword blue,
-// so they differ from jinja variables which share the plain propertyName tag
-const klipperConfigDarkHighlightStyle = HighlightStyle.define(
-    [
-        {
-            tag: tags.definition(tags.propertyName),
-            color: '#569cd6',
-        },
-    ],
-    { scope: klipperConfigLanguage, themeType: 'dark' }
-)
-
-const klipperConfigLightHighlightStyle = HighlightStyle.define(
-    [
-        {
-            tag: tags.definition(tags.propertyName),
-            color: '#0000ff',
-        },
-    ],
-    { scope: klipperConfigLanguage, themeType: 'light' }
-)
+import {
+    klipperConfigDarkHighlightStyle,
+    klipperConfigLightHighlightStyle,
+    yamlDarkHighlightStyle,
+    yamlLightHighlightStyle,
+} from '@/plugins/codemirror/highlightStyles'
 
 @Component
 export default class Codemirror extends Mixins(BaseMixin, ThemeMixin) {
