+++ src/plugins/codemirror/KlipperDocsTooltip.ts
