+++ src/plugins/codemirror/lezer/gcode.tokens.ts
